chore: refines Tooltip setup
