+++ src/components/Tooltip/index.module.scss
@@ -1,27 +1,29 @@
-@use '@scss/common' as *;
+@use '@scss/common.scss' as *;
 
 .tooltip {
-  @include label;
-  position: absolute;
-  bottom: 100%;
-  left: 50%;
-  z-index: 2;
-  transform: translate3d(-50%, -20%, 0);
-  padding: calc(var(--base) * .3) calc(var(--base) * .5);
-  color: var(--theme-text);
-  line-height: calc(var(--base) * .75);
-  font-weight: normal;
-  white-space: nowrap;
-  background-color: var(--theme-elevation-150);
-}
+  all: unset;
+  display: flex;
+  position: relative;
+  cursor: pointer;
+  border-radius: 3px;
 
-.caret {
-  position: absolute;
-  transform: translateX(-50%);
-  top: calc(100% - calc(var(--base) * .0625));
-  left: 50%;
-  height: 0;
-  width: 0;
-  border: 10px solid transparent;
-  border-top-color: var(--theme-elevation-150);
+  &:hover,
+  &:focus-visible {
+    background-color: var(--theme-elevation-200);
+  }
+
+  &:focus-visible {
+    @include outline;
+  }
+
+  .tip {
+    opacity: 0;
+    pointer-events: none;
+  }
+
+  &.show {
+    .tip {
+      opacity: 1;
+    }
+  }
 }
