+++ src/components/Tooltip/index.tsx
@@ -1,21 +1,67 @@
-import React from 'react'
+import * as React from 'react'
+
+import { TooltipContent } from '@components/Tooltip/TooltipContent'
 
 import classes from './index.module.scss'
 
-export type Props = {
+type TooltipProps = React.HTMLAttributes<HTMLButtonElement> & {
+  text: React.ReactNode
   children: React.ReactNode
-  className?: string
-}
+} & (
+    | {
+        isVisible?: never
+        setIsVisible?: never
+      }
+    | {
+        /**
+         * If this is set, the button will not manage its own state
+         */
+        isVisible: boolean
+        setIsVisible: (isActive: boolean) => void // eslint-disable-line no-unused-vars
+      }
+  )
 
-const Tooltip: React.FC<Props> = ({ children, className }) => {
-  const tooltipClasses = [classes.tooltip, className].filter(Boolean).join(' ')
+export const Tooltip: React.FC<TooltipProps> = ({
+  children,
+  className,
+  text,
+  onClick,
+  isVisible: isActive,
+  setIsVisible: setIsActive,
+}) => {
+  const [isVisibleInternal, setIsVisibleInternal] = React.useState(false)
+  const hoistControl = typeof setIsActive === 'function'
+  const show = hoistControl ? isActive : isVisibleInternal
+
+  const onFocusChange = React.useCallback(
+    (dir: string) => {
+      const nowActive = dir === 'enter'
+
+      if (hoistControl) {
+        setIsActive(nowActive)
+      } else {
+        setIsVisibleInternal(nowActive)
+      }
+    },
+    [setIsActive, hoistControl],
+  )
 
   return (
-    <aside className={tooltipClasses}>
+    <button
+      onFocus={() => onFocusChange('enter')}
+      onBlur={() => onFocusChange('leave')}
+      onMouseEnter={() => {
+        onFocusChange('enter')
+      }}
+      onMouseLeave={() => {
+        onFocusChange('leave')
+      }}
+      className={[classes.tooltip, show && classes.show, className].filter(Boolean).join(' ')}
+      type="button"
+      onClick={onClick}
+    >
       {children}
-      <span className={classes.caret} />
-    </aside>
+      <TooltipContent className={classes.tip}>{text}</TooltipContent>
+    </button>
   )
 }
-
-export default Tooltip
